docker/otel-elastic-test-ingest/requirements.txt
opentelemetry-distro[otlp]